Implement blockchain CG for older snapshots and blocks
Because disks are not infinite, we want to occasionally remove some
older blocks and snapshots so we don't fill the disk entirely. This
change adds an environment variable, BLOCKCHAIN_GC_BYTES that allows the
system to signal to the blockchain it needs to reclaim some space.

If this environment variable is set, the blockchain will kick off a
background process that, starting at the oldest blocks, will delete
blocks and snapshots at that height until it has deleted at least as
many bytes as was requested. Note that this will temporarily increase
disk usage because compaction will need to happen before rocksdb can
actually reclaim space.

When using this feature it's probably important that you remove *more*
than is your threshold for triggering GC. For example if you want to GC
when the disk is 80% full, you should probably tell the blockchain to
delete enough bytes to get you back to 75% so you have some hystersis.

The last 100k of blocks and snapshots are protected from GC.